input/keyboard: handle event->group_update
When a keyboard is added to or removed from a wlr_keyboard_group, a key
event will be emitted for all keys that only pressed for that keyboard
in the group. This allows for compositors to update their internal
state. However because these keys were already down, bindings should not
be executed and surfaces should not receive a key event.