Render active child last in linear tiling containers
This makes it so that if the "active" child has subsurfaces that extend
beyond the bounding box of the view, those are drawn above other
children. The idea is that you can interact with these subsurfaces
(such as GTK popovers) as long as the view is focused (or at least was
the last focused child in the container).

Relates to #4924