config/output: reconfigure on commit failure
If the wlr_output_commit fails and the output was previously enabled, we
re-enable the output. If the output was previously enabled, then it was
also configured. Without reenabling or reconfiguring the output, the
output is left in an invalid state. This adds the missing reconfiguring
step.

This also adds safe guards in output_disable to prevent removing an
output index that is not in the list.