Fix swaynag off-by-one issues in button positions
Button positions are off in both the x and y directions.

This can be easily seen if we set the thickness of button borders
to 1 and set all the other margins, paddings, border thicknesses
and button gaps to 0.

Then, in the y direction, there is an one-pixel gap between the top
of swaynag and the top button borders, while the bottom button borders
are invisible. In the x direction, the right border of the dismiss
button is invisible, while the borders of adjacent buttons overlap
by one pixel.

Fixed by removing the explicit one-pixel offsets in the assignment
statements for button->x and button-y in the render_button function.