Add for_window --always
This allows a for_window command to trigger multiple times on the same window;
the default is to only run the command once (when it first matches).

This can be useful to apply rules to applications like VLC that hide and show a
window without re-creating it (when using the taskbar icon to hide/show).