tree/container: use L_NONE for parent layout of floating container
The parent of a floating container is technically the workspace, but it
is incorrect to say that the parent layout of a floating container is
that of the workspace.

For instance, if the workspace layout is tabbed, a floating container
should not (and does not) show tab decorations. However, prior to this
commit, querying its parent layout would nonetheless return `L_TABBED`.

Fixes #5782.