Revert removal of raise_floating directive
This reverts commit b9b1b0e5662f8e3dbbfa9bb09dd4f69aac9c2db0 and
therefore pull request #2888. The reason being that with some panels
that list windows (such as wf-panel), focusing on a window won't raise
it up front, which is weird.

This pull request also sets the default value for raise_floating to
false, however, to not break existing set ups that expect the old
behavior.